apps/books/views.py
from rest_framework.generics import ListCreateAPIView
from apps.books.models import Book
from apps.books.serializers import BookSerializer


class BookListCreateAPIView(ListCreateAPIView):
    queryset = Book.objects.all()
    serializer_class = BookSerializer

